+++ VideoEngineController/RenderingBuffer.h
@@ -17,8 +17,8 @@ public:
 	CRenderingBuffer();
 	~CRenderingBuffer();
 
-	int Queue(int nFrameNumber, unsigned char *ucaDecodedVideoFrameData, int nLength, long long llCaptureTimeDifference, int nVideoHeight, int nVideoWidth, int nOrientation, int nInsetHeight, int nInsetWidth);
-	int DeQueue(int &rnFrameNumber, long long &rllCaptureTimeDifference, unsigned char *ucaDecodedVideoFrameData, int &rnVideoHeight, int &rnVideoWidth, int &rnTimeDifferenceInQueue, int &nOrientation, int &nInsetHeight, int &nInsetWidth);
+	int Queue(long long nFrameNumber, unsigned char *ucaDecodedVideoFrameData, int nLength, long long llCaptureTimeDifference, int nVideoHeight, int nVideoWidth, int nOrientation, int nInsetHeight, int nInsetWidth);
+	int DeQueue(long long &rnFrameNumber, long long &rllCaptureTimeDifference, unsigned char *ucaDecodedVideoFrameData, int &rnVideoHeight, int &rnVideoWidth, int &rnTimeDifferenceInQueue, int &nOrientation, int &nInsetHeight, int &nInsetWidth);
 	void IncreamentIndex(int &riIndex);
 	int GetQueueSize();
 	void ResetBuffer();
@@ -33,7 +33,7 @@ private:
 	unsigned char m_uc2aDecodedVideoDataBuffer[MAX_VIDEO_RENDERER_BUFFER_SIZE][MAX_VIDEO_RENDERER_FRAME_SIZE];
 
 	int m_naBufferDataLengths[MAX_VIDEO_RENDERER_BUFFER_SIZE];
-	int m_naBufferFrameNumbers[MAX_VIDEO_RENDERER_BUFFER_SIZE];
+	long long m_naBufferFrameNumbers[MAX_VIDEO_RENDERER_BUFFER_SIZE];
 	int m_naBufferVideoHeights[MAX_VIDEO_RENDERER_BUFFER_SIZE];
 	int m_naBufferVideoWidths[MAX_VIDEO_RENDERER_BUFFER_SIZE];
 	int m_naBufferVideoOrientations[MAX_VIDEO_RENDERER_BUFFER_SIZE];
