+++ VideoEngineController/VideoDecodingThreadOfChannel.h
@@ -56,7 +56,7 @@ namespace MediaSDK
 
 		void SetCallFPS(int nFPS);
 
-		int DecodeAndSendToClient(unsigned char *in_data, unsigned int frameSize, int nFramNumber, unsigned int nTimeStampDiff, int nOrientation, int nInsetHeight, int nInsetWidth);
+		int DecodeAndSendToClient(unsigned char *in_data, unsigned int frameSize, long long nFramNumber, long long nTimeStampDiff, int nOrientation, int nInsetHeight, int nInsetWidth);
 
 	private:
